refactor: Move parser.svelte under parser
