+++ src/lib/ChatScreens/BackgroundDom.svelte
@@ -1,5 +1,5 @@
 <script lang="ts">
-    import { ParseMarkdown, risuChatParser } from "src/ts/parser.svelte";
+    import { ParseMarkdown, risuChatParser } from "src/ts/parser/parser.svelte";
     import { type character, type groupChat } from "src/ts/storage/database.svelte";
     import { DBState } from 'src/ts/stores.svelte';
     import { moduleBackgroundEmbedding, ReloadGUIPointer, selIdState } from "src/ts/stores.svelte";
