+++ src/lib/ChatScreens/Suggestion.svelte
@@ -10,7 +10,7 @@
     import { language } from "src/lang";
     import { getUserName, replacePlaceholders } from "../../ts/util";
     import { onDestroy } from 'svelte';
-    import { ParseMarkdown } from "src/ts/parser.svelte";
+    import { ParseMarkdown } from "src/ts/parser/parser.svelte";
     import {defaultAutoSuggestPrompt} from "../../ts/storage/defaultPrompts.js";
 
     interface Props {
