+++ src/lib/Playground/PlaygroundParser.svelte
@@ -1,5 +1,5 @@
 <script lang="ts">
-    import { ParseMarkdown } from "src/ts/parser.svelte";
+    import { ParseMarkdown } from "src/ts/parser/parser.svelte";
     import TextAreaInput from "../UI/GUI/TextAreaInput.svelte";
     let input = $state("");
     let output = $state("");
