+++ src/lib/Playground/PlaygroundSyntax.svelte
@@ -1,6 +1,6 @@
 <script lang="ts">
     import TextAreaInput from "../UI/GUI/TextAreaInput.svelte";
-    import { risuChatParser } from 'src/ts/parser.svelte';
+    import { risuChatParser } from 'src/ts/parser/parser.svelte';
     import { language } from 'src/lang';
     import { sleep } from 'src/ts/util';
     let input = $state("");
