+++ src/lib/UI/GUI/MultiLangDisplay.svelte
@@ -1,6 +1,6 @@
 <script lang="ts">
     import { ColorSchemeTypeStore } from "src/ts/gui/colorscheme";
-    import { ParseMarkdown } from "src/ts/parser.svelte";
+    import { ParseMarkdown } from "src/ts/parser/parser.svelte";
     import { parseMultilangString, toLangName } from "src/ts/util";
 
     interface Props {
