+++ src/ts/cbs.ts
@@ -1,5 +1,5 @@
 import type { Database, character, loreBook } from './storage/database.svelte';
-import type { CbsConditions } from './parser.svelte';
+import type { CbsConditions } from './parser/parser.svelte';
 import type { RisuModule } from './process/modules';
 import type { LLMModel } from './model/modellist';
 import { get } from 'svelte/store';
