+++ src/ts/process/templates/jsonSchema.ts
@@ -1,4 +1,4 @@
-import { risuChatParser } from "src/ts/parser.svelte"
+import { risuChatParser } from "src/ts/parser/parser.svelte"
 import { getDatabase } from "src/ts/storage/database.svelte"
 import { jsonOutputTrimmer } from "src/ts/util"
